+++ src/JokeFactory.php
@@ -4,16 +4,35 @@ declare(strict_types=1);
 
 namespace Gopibabu\Jokes;
 
+
+use GuzzleHttp\Client;
+use GuzzleHttp\Exception\GuzzleException;
+
 /**
  * Class JokeFactory.
  */
 class JokeFactory
 {
+    const API_ENDPOINT = 'http://api.icndb.com/jokes/random';
+
+    protected $client;
+
     /**
-     * @return string
+     * @param Client $client
+     */
+    public function __construct(Client $client = null){
+        $this->client = $client ?: new Client();
+    }
+
+    /**
+     * @return mixed
+     * @throws GuzzleException
      */
     public function generateJoke()
     {
-        return 'This is a random joke';
+        $response = $this->client->get(self::API_ENDPOINT);
+        $joke =  json_decode($response->getBody()->getContents());
+
+        return $joke->value->joke;
     }
 }