+++ tests/JokeFactoryTest.php
@@ -1,10 +1,13 @@
 <?php
-
 declare(strict_types=1);
 
 namespace Gopibabu\Jokes\Tests;
 
 use Gopibabu\Jokes\JokeFactory;
+use GuzzleHttp\Client;
+use GuzzleHttp\Handler\MockHandler;
+use GuzzleHttp\HandlerStack;
+use GuzzleHttp\Psr7\Response;
 use PHPUnit\Framework\TestCase;
 
 class JokeFactoryTest extends TestCase
@@ -12,7 +15,16 @@ class JokeFactoryTest extends TestCase
     /** @test */
     public function returnsRandomJoke()
     {
-        $joke = new JokeFactory();
-        $this->assertSame('This is a random joke', $joke->generateJoke());
+        $mock = new MockHandler(
+            [
+                new Response(200, [], '{ "type": "success", "value": { "id": 356, "joke": "We live in an expanding universe. All of it is trying to get away from Chuck Norris.", "categories": [] } }')
+            ]
+        );
+
+        $handlerStack = HandlerStack::create($mock);
+        $client = new Client(['handler' => $handlerStack]);
+
+        $joke = new JokeFactory($client);
+        $this->assertSame("We live in an expanding universe. All of it is trying to get away from Chuck Norris.", $joke->generateJoke());
     }
 }