+++ changelog.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - Added Changelog file to view changes added for different releases.
 - Added .gitattributes file to ignore files and folders while exporting this package
 - Added Issue template to create issues in Github.
+- Added endpoint to generate random joke.
 
 ## [1.3.2] - 10-15-2020
 ### Updated